repaired int overflow , hopefully , and created test input

+++ 03-set/rp.cpp
@@ -1,27 +1,26 @@
 #include <iostream>
 #include <vector>
 #include <algorithm>
-
-
+#include <cstdint>
 
 
 int main()
 {
 
-        int n, m;
+        int64_t n, m;
         std::cin >> n >> m ;
-        std::vector <int> can_sizes;
+        std::vector <int64_t> can_sizes;
         can_sizes.reserve(n);
-        std::vector <int> colors_amout;
+        std::vector <int64_t> colors_amout;
         colors_amout.reserve(m);
         
-        for (int i = 0 ; i < n ; i++) {
-                int x;
+        for (int64_t i = 0 ; i < n ; i++) {
+                int64_t x;
                 std::cin >> x;
                 can_sizes.push_back(x);
         }
-        for (int i = 0 ; i < m ; i++) {
-                int x;
+        for (int64_t i = 0 ; i < m ; i++) {
+                int64_t x;
                 std::cin >> x;
                 colors_amout.push_back(x);
         }
@@ -29,7 +28,7 @@ int main()
         std::sort(colors_amout.begin(), colors_amout.end());
 
 
-        int wasted = 0;
+        int64_t wasted = 0;
         auto can = can_sizes.begin();
         for (auto color: colors_amout){
                 while ( color > *can ){
